+++ Assets/Scripts/Player/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d367910012babd043ae6d46371a52d8d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
